Switch to rspec for testing.
Use rspec for testing instead of test/unit for more
flexibility and ability to use existing gems to help
with different kinds of testing.

Include the vcr and webmock gems with rspec to start
using mocks and fixtures to test the outbound API
calls to Nike without needing to always have a valid
internet connection. This commit only tests the
list_activities endpoint.

- Add vcr and webmock dependencies to Gemfile for bundler.
- Include configuration of vcr in spec_helper.
- Add fixtures for quick testing of endpoints.
- Begin testing of endpoints using vcr and webmock.
- Setup task in Rakefile for running rspec tests.
- Remove test/unit tests and Rakefile task.
- Translate existing tests into rspec.
- Include rspec as dependency with Bundler.
- Create spec_helper to include the NikePlusGem root file.